backend/pyproject.toml
[project]
name = "mathfluent-backend"
version = "0.1.0"
description = "Backend API for MathFluent application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "openai>=1.65.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Explicitly tell build system where to find the packages
[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88 
